+++ src/player.h
@@ -9,7 +9,7 @@ typedef struct {
     Rack *rack;
     int last_note[TRACKS_PER_PATTERN];
     int last_instrument[TRACKS_PER_PATTERN];
-    int pattern_pos;
+    Uint8 pattern_pos;
     int substep;
     Uint32 tempo;
     SDL_TimerID timer;
